// src/app/core/interfaces/registeredUser.ts
export interface RegisteredUserEvents {

    id: string;
    ufname: string;
    ulname: string;
    uemail: string;
    uphone: string | null;
    uage: string;
    uTicketType: string;
    uCity: string;
    uTicketCount: string;
    eventName: string;
    eventTime: string;
    eventVenue: string;
    eventImage: string;
    eventLanguage: string;
    paymentStatus: string;
}